convertCsv is not templated anymore
moved implementation back into *.cpp
