+++ csv2xlslib/parseCsvFile.cpp
@@ -1,10 +1,10 @@
+#include "convertCsv.h"
 #include "parseCsvFile.hpp"
 #include "XlsWorkBook.hpp"
 #include "conversion.h"
 #include "readHeadLine.hpp"
 #include "Helpers.h"
 #include <sstream>
-#include "convertCsv.h"
 #include "readBuffer.h"
 #include "funcomp.h"
